requirements.txt
strands-agents
strands-agents-tools
mcp
streamlit
boto3
diagrams
wheel
setuptools
